refactor(tooling): split materiality context logic
